swdc_green_iot_hackathon/swdc_flask/requirements.txt
aniso8601==8.0.0
astroid==2.4.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
paho-mqtt==1.5.0
pycparser==2.20
pylint==2.5.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.23.0
rope==0.17.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
